+++ src/jk_testing/utils/__init__.py
@@ -1,10 +1,9 @@
 ﻿
 
-__version__ = "0.2021.3.25"
+__version__ = "0.2021.5.25"
 
 
 
 from .SimpleTextLineOutputBuffer import SimpleTextLineOutputBuffer
 from .OutputComparison import OutputComparison
-