+++ HISTORY.md
@@ -3,3 +3,6 @@
 
 * 2021-03-25
 	* Added 'utils' subpackage with utility classes
+
+* 2021-05-25
+	* Added isRegExMatch() and isNotNoneOrEmpty() to Assert